backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
# Database and migration dependencies
sqlalchemy>=1.4
psycopg2-binary
alembic